Title Screen
Finished a majority of the title screen. Added water and a beach and some textures to go with them and also made the water move a little. I also added a boat but right now its just a cube. I also added buttons but they are also placeholder for right now